+++ script.artistslideshow/addon.xml
@@ -1,4 +1,4 @@
-<addon id="script.artistslideshow" name="Artist Slideshow" provider-name="pkscout,ronie" version="3.0.0">
+<addon id="script.artistslideshow" name="Artist Slideshow" provider-name="pkscout,ronie" version="3.1.0">
 	<requires>
 		<import addon="xbmc.python" version="2.26.0" />
 		<import addon="xbmc.addon" version="17.9.0" />
@@ -12,19 +12,24 @@
 	</extension>
 	<extension point="xbmc.addon.metadata">
 		<news>
-v.3.0.0
-- rewrite to use a single image control instead of multi-image control in skin
-- AS can now store and show music in Kodi 18 music artist folder
-- AS now shows the Kodi artist fanart (if available) before anything else (with option to disable)
-- added option to include Kodi album fanart in image list (defaults to False)
-- added routine in settings to move images to Kodi artist information folder
-- added option to disable extranfanart folder
-- added option to control time between slide transitions
-- AS can now store images and information files in separate directory structure
-- AS now deletes empty artist information and artwork folders
-- converted language files to current format for forward compatibility
-- removed option to have a pop up displayed when downloading artwork
-- Python2/3 compatible
+v.3.1.0
+- In daemon mode you can now change AS settings without having to restart Kodi
+- added back option to show notifications when downloading new images
+- updated service plugins to use https for API calls
+- added options to change sleep time for main and slideshow threads
+- set slideshow thread to close if no music playing
+- added option to search stream titles more aggressively for an artist name
+- fix for crash if artist image name has an extended character in it
+- fix for crash when trying to get last fanart number from Kodi Artist Information Folder
+- fix for artist images not changing when using a radio stream
+- fix for theaudiodb.com plugin not downloading all three images for an artist
+- fix for unicode errors causing cache trimming not to work
+- fix for non-images getting added to image list and showing a black screen
+- fix for high CPU usage when using AS in daemon mode and no music playing
+- fix for issue where slideshow might not start when in daemon mode
+- fix for runaway loop if there is only one image in an artist's slideshow
+- fix for AS taking too long to quit on Kodi exit when in daemon mode
+- fix for Setting Type Error in Kodi logs
 		</news>
 		<assets>
         	<icon>icon.png</icon>
