+++ script.artistslideshow/resources/images/update-slide/notice.png
