+++ script.artistslideshow/resources/lib/fileops.py
