+++ script.artistslideshow/resources/lib/url.py
