+++ script.artistslideshow/resources/lib/xlogger.py
