+++ script.artistslideshow/resources/plugins/fanarttv.py
@@ -2,8 +2,8 @@
 
 import base64, os, time, random
 from kodi_six import xbmcvfs
-from resources.common.url import URL
-from resources.common.fileops import readFile, writeFile, deleteFile, checkPath
+from resources.lib.url import URL
+from resources.lib.fileops import readFile, writeFile, deleteFile, checkPath
 from kodi_six.utils import py2_encode
 import json as _json
 try:
@@ -18,7 +18,7 @@ except AttributeError:
 class objectConfig( object ):
     def __init__( self ):
         secsinweek = int( 7*24*60*60 )
-        self.URL = 'http://webservice.fanart.tv/v3/music/'
+        self.URL = 'https://webservice.fanart.tv/v3/music/'
         self.FILENAME = 'fanarttvartistimages.nfo'
         self.CACHETIMEFILENAME = 'fanarttvcachetime.nfo'
         self.HASCLIENTKEY = False
@@ -28,7 +28,7 @@ class objectConfig( object ):
         self.CACHEEXPIRE['high'] = int( 4*secsinweek )
         self.CACHEEXPIREWITHCLIENTKEY = int( 2*secsinweek )
         self.CACHEEXPIREWITHDONATION = int( secsinweek/7 )
-        self.loglines = []
+        self.LOGLINES = []
         self.JSONURL = URL( 'json' )
 
 
@@ -37,7 +37,7 @@ class objectConfig( object ):
 
 
     def getImageList( self, img_params ):
-        self.loglines = []
+        self.LOGLINES = []
         url_params = {}
         images = []
         filepath = os.path.join( img_params.get( 'infodir', '' ), self.FILENAME )
@@ -47,7 +47,7 @@ class objectConfig( object ):
         if img_params.get( 'clientapikey', False ):
             self.HASCLIENTKEY = True
             url_params['client_key'] = img_params.get( 'clientapikey', '' )
-            if img_params.get( 'donate' ) == 'true':
+            if img_params.get( 'donated', False ):
                 self.HASDONATION = True
                 self.CACHEEXPIRE['low'] = self.CACHEEXPIREWITHDONATION
                 self.CACHEEXPIRE['high'] = self.CACHEEXPIREWITHDONATION
@@ -57,27 +57,27 @@ class objectConfig( object ):
         json_data = self._get_data( filepath, cachefilepath, url, url_params )
         if json_data:
             image_list = json_data.get( 'artistbackground', [] )
-            if img_params.get( 'getall', 'false' ) == 'true':
+            if img_params.get( 'getall', False ):
                 image_list.extend( json_data.get( 'artistthumb', [] ) )
             for image in image_list:
                 url = image.get( 'url', '' )
                 if url:
                     images.append( url )
-        return images, self.loglines
+        return images, self.LOGLINES
 
 
     def _get_cache_time( self, cachefilepath ):
         rawdata = ''
-        self.loglines.append( 'getting the cache timeout information for fanarttv' )
+        self.LOGLINES.append( 'getting the cache timeout information for fanarttv' )
         exists, cloglines = checkPath( cachefilepath, False )
-        self.loglines.extend( cloglines )
+        self.LOGLINES.extend( cloglines )
         if exists:
             success = True
         else:
             success = self._put_cache_time( cachefilepath )
         if success:
             rloglines, rawdata = readFile( cachefilepath )
-            self.loglines.extend( rloglines )
+            self.LOGLINES.extend( rloglines )
         try:
             cachetime = int( rawdata )
         except ValueError:
@@ -95,21 +95,21 @@ class objectConfig( object ):
         json_data = ''
         if self._update_cache( filepath, cachefilepath ):
             success, uloglines, json_data = self.JSONURL.Get( url, params=url_params )
-            self.loglines.extend( uloglines )
+            self.LOGLINES.extend( uloglines )
             if success:
                 success, wloglines = writeFile( py2_encode( _json.dumps( json_data ) ), filepath )
-                self.loglines.extend( wloglines )
+                self.LOGLINES.extend( wloglines )
         exists, cloglines = checkPath( filepath, False )
-        self.loglines.extend( cloglines )
+        self.LOGLINES.extend( cloglines )
         if exists:
             rloglines, rawdata = readFile( filepath )
-            self.loglines.extend( rloglines )
+            self.LOGLINES.extend( rloglines )
             try:
                 json_data = _json.loads( rawdata )
             except ValueError:
                 success, dloglines = deleteFile( filepath )
-                self.loglines.extend( dloglines )
-                self.loglines.append( 'Deleted old cache file. New file will be download on next run.' )
+                self.LOGLINES.extend( dloglines )
+                self.LOGLINES.append( 'Deleted old cache file. New file will be download on next run.' )
                 json_data = ''
         return json_data
 
@@ -117,21 +117,21 @@ class objectConfig( object ):
     def _put_cache_time( self, cachefilepath ):
         cachetime = random.randint( self.CACHEEXPIRE.get( 'low' ), self.CACHEEXPIRE.get( 'high' ) )
         success, wloglines = writeFile( str( cachetime ), cachefilepath )
-        self.loglines.append( wloglines)
+        self.LOGLINES.append( wloglines)
         return success
 
 
     def _update_cache( self, filepath, cachefilepath ):
         exists, cloglines = checkPath( filepath, False )
-        self.loglines.extend( cloglines )
+        self.LOGLINES.extend( cloglines )
         if exists:
             st = xbmcvfs.Stat( filepath )
             if time.time() - st.st_mtime() < self._get_cache_time( cachefilepath ):
-                self.loglines.append( 'cached artist info found for fanarttv' )
+                self.LOGLINES.append( 'cached artist info found for fanarttv' )
                 return False
             else:
-                self.loglines.append( 'outdated cached artist info found for fanarttv' )
+                self.LOGLINES.append( 'outdated cached artist info found for fanarttv' )
                 return self._put_cache_time( cachefilepath )
         else:
-            self.loglines.append( 'no fanarttv cachetime file found, creating it' )
+            self.LOGLINES.append( 'no fanarttv cachetime file found, creating it' )
             return self._put_cache_time( cachefilepath )
