+++ script.artistslideshow/resources/plugins/lastfm.py
@@ -2,8 +2,8 @@
 
 import base64, os, time, random
 import defusedxml.ElementTree as _xmltree
-from resources.common.url import URL
-from resources.common.fileops import readFile, writeFile, checkPath
+from resources.lib.url import URL
+from resources.lib.fileops import readFile, writeFile, checkPath
 from kodi_six.utils import py2_encode
 try:
     from . import lastfm_info as settings
@@ -24,7 +24,7 @@ class objectConfig( object ):
                             'api_key':base64.b64decode(clowncar.encode('ascii')).decode('ascii'), 'method':'artist.getTopAlbums'}
         self.SIMILARPARAMS = {'autocorrect':'1',
                               'api_key':base64.b64decode(clowncar.encode('ascii')).decode('ascii'), 'limit':'50', 'method':'artist.getSimilar'}
-        self.URL = 'http://ws.audioscrobbler.com/2.0/'
+        self.URL = 'https://ws.audioscrobbler.com/2.0/'
         self.BIOFILENAME = 'lastfmartistbio.nfo'
         self.ALBUMFILENAME = 'lastfmartistalbums.nfo'
         self.SIMILARFILENAME = 'lastfmartistsimilar.nfo'
